move ModelMapper to a separate configuration class
