+++ src/main/java/com/rexdelf/issuetrackerapp/IssueTrackerAppApplication.java
@@ -7,10 +7,6 @@ import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 public class IssueTrackerAppApplication {
-  @Bean
-  public ModelMapper modelMapper(){
-    return new ModelMapper();
-  }
 
   public static void main(String[] args) {
     SpringApplication.run(IssueTrackerAppApplication.class, args);
